Refactoring DSL attribute file path
