+++ lib/netomox/dsl/link_attr/mddo.rb
