+++ lib/netomox/dsl/link_attr/rfc.rb
