+++ lib/netomox/dsl/network_attr/mddo.rb
