+++ lib/netomox/dsl/network_attr/rfc.rb
