+++ lib/netomox/dsl/node_attr/mddo_static_route.rb
