+++ lib/netomox/dsl/node_attr/rfc_prefix.rb
