+++ lib/netomox/dsl/tp_attr/ops.rb
